Return GitHub user's current rate limits

Reports the current rate limit status for the authenticated user, either pulls this information from a previous successful request or directly from the GitHub API.

gh_rate_limit(
  response = NULL,
  .token = NULL,
  .api_url = NULL,
  .send_headers = NULL
)

Arguments

response
gh_response object from a previous gh call, rate limit values are determined from values in the response header. Optional argument, if missing a call to "GET /rate_limit" will be made.

.token
Authentication token. Defaults to GITHUB_PAT or GITHUB_TOKEN environment variables, in this order if any is set. See gh_token() if you need more flexibility, e.g. different tokens for different GitHub Enterprise deployments.

.api_url
Github API url (default: https://api.github.com). Used if endpoint just contains a path. Defaults to GITHUB_API_URL environment variable if set.

.send_headers
Named character vector of header field values (except Authorization, which is handled via .token). This can be used to override or augment the default User-Agent header: "https://github.com/r-lib/gh".

Value

A list object containing the overall limit, remaining limit, and the limit reset time.

Details

Further details on GitHub's API rate limit policies are available at https://docs.github.com/v3/#rate-limiting.