Compute the suffix that gh uses for GitHub API URL specific PATs

slugify_url() determines a suffix from a URL and this suffix is used to construct the name of an environment variable that holds the PAT for a specific GitHub URL. This is mostly relevant to people using GitHub Enterprise. slugify_url() processes the API URL like so:

Extract the host name, i.e. drop both the protocol and any path

Substitute "github.com" for "api.github.com"

Replace special characters with underscores

Convert to ALL CAPS

This suffix is then added to GITHUB_PAT_ to form the name of an environment variable. It's probably easiest to just look at some examples.

# both give same result
slugify_url("https://api.github.com")

## [1] "GITHUB_COM"

slugify_url("https://github.com")

## [1] "GITHUB_COM"

# an instance of GitHub Enterprise
# both give same result
slugify_url("https://github.acme.com")

## [1] "GITHUB_ACME_COM"

slugify_url("https://github.acme.com/api/v3")

## [1] "GITHUB_ACME_COM"

slugify_url(url)

Arguments

url
Character vector of HTTP/HTTPS URLs. They don't have to be in the API-specific form, although they can be.

Value

Character vector of suffixes.

See also

Examples

# main github.com site
slugify_url("https://api.github.com")

#> [1] "GITHUB_COM"

slugify_url("https://github.com")

#> [1] "GITHUB_COM"

# an instance of GitHub Enterprise
slugify_url("https://github.acme.com")

#> [1] "GITHUB_ACME_COM"

slugify_url("https://github.acme.com/api/v3")

#> [1] "GITHUB_ACME_COM"