gh_next {gh}	R Documentation

Get the next, previous, first or last page of results

Description

Get the next, previous, first or last page of results

Usage

gh_next(gh_response)

gh_prev(gh_response)

gh_first(gh_response)

gh_last(gh_response)

Arguments

gh_response
An object returned by a gh() call.

Details

Note that these are not always defined. E.g. if the first page was queried (the default), then there are no first and previous pages defined. If there is no next page, then there is no next page defined, etc.

If the requested page does not exist, an error is thrown.

Value

Answer from the API.

Examples

## Not run: 
x <- gh("/users")
sapply(x, "[[", "login")
x2 <- gh_next(x)
sapply(x2, "[[", "login")

## End(Not run)

[Package gh version 1.0.1 Index]