Get the next, previous, first or last page of results

Get the next, previous, first or last page of results

Usage

gh_next(gh_response)

gh_prev(gh_response)

gh_first(gh_response)

gh_last(gh_response)

Arguments

gh_response

An object returned by a gh() call.

Value

Answer from the API.

Details

Note that these are not always defined. E.g. if the first page was queried (the default), then there are no first and previous pages defined. If there is no next page, then there is no next page defined, etc.

If the requested page does not exist, an error is thrown.

See also

Examples

x <- gh("/users")
vapply(x, "[[", character(1), "login")
#>  [1] "mojombo"      "defunkt"      "pjhyett"      "wycats"      
#>  [5] "ezmobius"     "ivey"         "evanphx"      "vanpelt"     
#>  [9] "wayneeseguin" "brynary"      "kevinclark"   "technoweenie"
#> [13] "macournoyer"  "takeo"        "caged"        "topfunky"    
#> [17] "anotherjesse" "roland"       "lukas"        "fanvsfan"    
#> [21] "tomtt"        "railsjitsu"   "nitay"        "kevwil"      
#> [25] "KirinDave"    "jamesgolick"  "atmos"        "errfree"     
#> [29] "mojodna"      "bmizerany"   
x2 <- gh_next(x)
vapply(x2, "[[", character(1), "login")
#>  [1] "jnewland"    "joshknowles" "hornbeck"    "jwhitmire"   "elbowdonkey"
#>  [6] "reinh"       "knzai"       "bs"          "rsanheim"    "schacon"    
#> [11] "uggedal"     "bruce"       "sam"         "mmower"      "abhay"      
#> [16] "rabble"      "benburkert"  "indirect"    "fearoffish"  "ry"         
#> [21] "engineyard"  "jsierles"    "tweibley"    "peimei"      "brixen"     
#> [26] "tmornini"    "outerim"     "daksis"      "sr"          "lifo"