Return the local user's GitHub Personal Access Token (PAT)

If gh can find a personal access token (PAT) via gh_token(), it includes the PAT in its requests. Some requests succeed without a PAT, but many require a PAT to prove the request is authorized by a specific GitHub user. A PAT also helps with rate limiting. If your gh use is more than casual, you want a PAT.

The PAT corresponding to api_url is searched for with a strategy that looks in one or more of these places:

"env": environment variable(s)

"git": Git credential store (requires the credentials package)

"key": OS-level keychain (requires the keyring package)

Details are in the Managing Personal Access Tokens vignette.

gh_token(api_url = NULL, strategy = NULL)

Arguments

api_url
GitHub API URL. Defaults to the GITHUB_API_URL environment variable, if set, and otherwise to https://api.github.com.

strategy
Where to look for a PAT. If specified, must be a comma-delimited string consisting of "env", "git", and/or "key". Examples: "env", "env,git", "key,git,env". gh searches for a PAT in these places, in this order.

By default, strategy is "env,git" if the credential package is available and "env" if it is not.

Value

A string of 40 hexadecimal digits, if a PAT is found, or the empty string, otherwise. For convenience, the return value has an S3 class in order to ensure that simple printing strategies don't reveal the entire PAT.

See also

Examples

if (FALSE) {
gh_token()

format(gh_token())

str(gh_token())
}